+++ src/Propulsion.Cosmos/EquinoxCosmosParser.fs
@@ -16,14 +16,6 @@ module EquinoxCosmosParser =
             tmp.SetPropertyValue("content", document)
             tmp.GetPropertyValue<'T>("content")
 
-    // TODO remove when using Equinox rc5, where Event isa IEvent
-    let (|PropulsionEvent|) (x: Equinox.Cosmos.Store.Event) =
-        { new Gardelloyd.IEvent<_> with
-            member __.EventType = x.c
-            member __.Data = x.d
-            member __.Meta = x.m
-            member __.Timestamp = x.t }
-
     /// Sanity check to determine whether the Document represents an `Equinox.Cosmos` >= 1.0 based batch
     let isEquinoxBatch (d : Document) = 
         d.GetPropertyValue "p" <> null && d.GetPropertyValue "i" <> null
@@ -31,7 +23,7 @@ module EquinoxCosmosParser =
 
     /// Enumerates the events represented within a batch
     let enumEquinoxCosmosEvents (batch : Equinox.Cosmos.Store.Batch) : StreamEvent<byte[]> seq =
-        batch.e |> Seq.mapi (fun offset (PropulsionEvent x) -> { stream = batch.p; index = batch.i + int64 offset; event = x })
+        batch.e |> Seq.mapi (fun offset x -> { stream = batch.p; index = batch.i + int64 offset; event = x })
 
     /// Collects all events with a Document [typically obtained via the CosmosDb ChangeFeed] that potentially represents an Equinox.Cosmos event-batch
     let enumStreamEvents (d : Document) : StreamEvent<byte[]> seq =