+++ tests/Propulsion.Tests/StreamStateTests.fs
@@ -7,7 +7,7 @@ open Xunit
 
 let canonicalTime = System.DateTimeOffset.UtcNow
 
-let mk p c : StreamSpan<string> = { index = p; events = [| for x in 0..c-1 -> Gardelloyd.Core.EventData.Create(p + int64 x |> string, null, timestamp=canonicalTime) |] }
+let mk p c : StreamSpan<string> = { index = p; events = [| for x in 0..c-1 -> FsCodec.Core.EventData.Create(p + int64 x |> string, null, timestamp=canonicalTime) |] }
 let mergeSpans = StreamSpan.merge
 let trimSpans = StreamSpan.dropBeforeIndex
 let is (xs : StreamSpan<string>[]) (res : StreamSpan<string>[]) =